refactoring of interface into the common interface.ts file and misc changes
