+++ packages/components/nodes/chatmodels/ChatOpenAI/FlowiseChatOpenAI.ts
@@ -1,8 +1,7 @@
 import type { ClientOptions } from 'openai'
 import { ChatOpenAI as LangchainChatOpenAI, OpenAIChatInput, LegacyOpenAIInput, AzureOpenAIInput } from '@langchain/openai'
 import { BaseChatModelParams } from '@langchain/core/language_models/chat_models'
-import { IMultiModalOption } from '../../../src'
-import { IVisionChatModal } from '../../../src/IVisionChatModal'
+import { IMultiModalOption, IVisionChatModal } from '../../../src'
 
 export class ChatOpenAI extends LangchainChatOpenAI implements IVisionChatModal {
     configuredModel: string
